# im-backoffice/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  errors.Response:
    properties:
      key:
        type: string
      message:
        type: string
    type: object
...
  warehouse.LogisticCenterDto:
    properties:
      logistic_center_id:
        type: string
      logistic_center_name:
        type: string
    type: object
  warehouse.WarehouseReq:
    properties:
      logistic_center_id:
        type: string
      warehouse_name:
        type: string
    required:
    - logistic_center_id
    - warehouse_name
    type: object
  warehouse.WarehouseRes:
    properties:
      logistic_centers:
        $ref: '#/definitions/warehouse.LogisticCenterDto'
      warehouse_id:
        type: string
      warehouse_name:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is an inventory management server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Inventory Management API
  version: "2.0"
paths:
...
  /warehouse:
    post:
      consumes:
      - application/json
      parameters:
      - description: Warehouse
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/warehouse.WarehouseReq'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/warehouse.WarehouseRes'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/errors.Response'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/errors.Response'
      security:
      - ApiKeyAuth: []
      summary: Create a new warehouse
      tags:
      - warehouse
  /warehouse/{warehouseId}:
    delete:
      parameters:
      - description: Warehouse ID
        in: path
        name: warehouseId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/errors.Response'
      security:
      - ApiKeyAuth: []
      summary: Create a new warehouse
      tags:
      - warehouse
    put:
      consumes:
      - application/json
      parameters:
      - description: Warehouse DTO
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/warehouse.WarehouseReq'
      - description: Warehouse ID
        in: path
        name: warehouseId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/warehouse.WarehouseRes'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/errors.Response'
      security:
      - ApiKeyAuth: []
      summary: Update the warehouse.
      tags:
      - warehouse
schemes:
- http
- https
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"